调整天勤支持机制
